Conflicts:
	app-editors/gvim/Manifest
	app-editors/gvim/gvim-7.2.303.ebuild
	app-vim/conque/Manifest
	app-vim/nerdcommenter/Manifest
	dev-python/docutils/Manifest
	dev-python/docutils/docutils-0.6.ebuild
	dev-python/ipython/Manifest
	dev-util/cmake/Manifest
	media-video/nvidia-settings/Manifest
	net-irc/weechat/weechat-0.2.6.1.ebuild
	net-misc/remmina-plugins/Manifest
	net-misc/remmina/Manifest
	profiles/package.mask/funtoo
	profiles/repo_name
	sci-libs/scipy/Manifest
	sci-visualization/veusz/Manifest
	sys-boot/grub-legacy/Manifest
	sys-fs/cryptsetup/Manifest
	x11-libs/qt-gui/Manifest
	x11-libs/qt-webkit/Manifest

+++ gnome-base/gnome-keyring/gnome-keyring-2.32.1.ebuild
@@ -1,6 +1,6 @@
-# Copyright 1999-2010 Gentoo Foundation
+# Copyright 1999-2011 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/gnome-base/gnome-keyring/gnome-keyring-2.32.1.ebuild,v 1.1 2010/11/01 16:18:00 eva Exp $
+# $Header: /var/cvsroot/gentoo-x86/gnome-base/gnome-keyring/gnome-keyring-2.32.1.ebuild,v 1.4 2011/01/02 21:32:23 mr_bones_ Exp $
 
 EAPI="3"
 GCONF_DEBUG="yes"
@@ -35,6 +35,10 @@ PDEPEND="gnome-base/libgnome-keyring"
 
 DOCS="AUTHORS ChangeLog NEWS README"
 
+# tests fail in several ways, they should be fixed in the next cycle (bug #340283),
+# revisit then.
+RESTRICT="test"
+
 pkg_setup() {
 	G2CONF="${G2CONF}
 		$(use_enable debug)
@@ -63,8 +67,10 @@ src_prepare() {
 
 src_install() {
 	gnome2_src_install
-	find "${ED}"/$(get_libdir)/security -name "*.la" -delete \
-		|| die "la file removal failed"
+	if use pam; then
+		find "${ED}"/$(get_libdir)/security -name "*.la" -delete \
+			|| die "la file removal failed"
+	fi
 }
 
 src_test() {
